requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.8.3
protobuf>=3.20.2,<6.0.0
dotenv==1.0.0